+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This log documents significant changes for each release.  This project follows
 ### Added
 - String manipulation functions: upper(), lower(), toChars().
 ### Fixed
-- String manipulation functions did not be properly returning an empty
+- String manipulation functions did not properly return an empty
   collection when the input collection is empty.
   
 ## [2.7.4] - 2021-03-12